# docs/swagger.yaml
definitions:
  controller.CommonError:
    properties:
      error:
        type: string
    type: object
  controller.CommonResponse:
    properties:
      message:
        type: string
    type: object
  controller.CreateUserRequest:
    properties:
      age:
        type: integer
      email:
        type: string
      user_name:
        type: string
    type: object
  models.User:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      user_name:
        type: string
    type: object
info:
  contact:
    email: help@socialetwork.com
    name: SocialNetwork API Service
    url: https://socialnetwork.com
  title: SocialNetwork API
paths:
  /users:
    get:
      description: Get list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Get all users
      tags:
      - Users
    post:
      description: Create new user and add to database
      parameters:
      - description: new user data
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/controller.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Create user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Delete user's data by ID
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Delete user by ID
      tags:
      - Users
    get:
      description: Get user's data by ID
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Get user by ID
      tags:
      - Users
    put:
      description: Update user's data by ID
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      - description: updated user data
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/controller.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Update user by ID
      tags:
      - Users
swagger: "2.0"